remove hardcoded prefix "Quelle: " from figsrcinfo.lua filter
This way the filter is language independed, i.e. it can be
used for presentations in German as well as in English or ...

It would be convenient to automatically add some prefix like "Quelle: "
to the caption, but then we would need some localisation mechanism.
May be sometime in the future :)