add filter for solution spaces
fenced divs with class "solution" and attribute "length"
will be translated to a solution box (see exam class)
with a blue/gray marker on the left side. this field will
be of length "length" and appear empty in the exam sheet.
in the solution sheet the content will appear. note: due
to the exam class the length of this field will fit the
content (i.e. answer) in the solution sheet ... so add
some `\vspace{}`, otherwise the remaining content will
not appear on the same page as in the exam sheet!

when used without attributes, i.e. without attribute `length`,
the content will appear in both the exam and the solution with
a blue/gray marker on the left side. since a minipage is used
behind, markdown tables cannot be used (pandoc uses `longtable`,
which cannot be used within a minipage). the workaraound would
be a custom filter which translates markdown tables to simple
latex tables ...