add filter for `[...]{.origin}` span
using the title text of an pandoc image we can only use
plain text. however, sometimes we need a little formatting
or even urls. for *inline images* this span can be used. it
need to be in the same line as the inline image. for *figures*
with a real caption one could always extend the caption ...