CSS: deaktivate dropdown menu in nav bar for small screens
Small screens (like the resolution defined here) are mainly found in
tablet devices with touch screens. On touch screens, however, open
dropdowns cannot be closed in pure CSS. So this change improves the
situation for tablet devices.

On a desktop where a mouse is available, this change reduces the
convenience a little. However, rough navigation is available via
the sections (which are still displayed in the navigation bar).

So that seems like a good compromise. The alternative would be to
forego pure CSS and work with Javascript instead.