fix handling of cbox class
A cbox is actually a div instead of a span - the span used
earlier had to appear in a separate paragraph for this to work.
A div or `\ begin {center}` was then created around it in the
filter ... In addition, the linebreak in narrow browser windows
(e.g. mobile phone) did not work.

With this change, a cbox is treated as a div. This simplifies
the CSS definition and the filter for wrapping for HTML
generation is no longer necessary. For TeX a new environment
has to be created instead of a command. There is also a new
filter that issues a warning when the old cbox span is used.

Note: Until now the cbox was designed for a one-liner, and
the box was only as wide as the text was long. Treatment as
a div now means that the width must be artificially limited.
In HTML, 40% text width and in TeX slides 60% text width
seem to be a good measure.