fix navigation for really small screens
On very small screens (e.g. cell phones), a navigation menu
fixed at the top is less useful, as it takes up too much space.

With a pure CSS solution, you would have to generate one (or
two) buttons in the HTML code that are only visible on narrow
screens and make the navigation menu visible and invisible
again when you click. This requires intervention in the creation
of the HTML handout with Pandoc (e.g. a filter or a mix-in HTML
snippet).

As an interim solution, the menu is placed as a normal element
at the beginning of the document, i.e. it scrolls off the screen
over time. However, this is still useful, as you can quickly get
an overview of the document and jump to a later section if necessary.