add home button to handout (html)
to allow easier navigation within the html handout to a 'home document' you can add a
variable 'home' to the metadata (e.g. in the YAML header) containing a string with the
URI of this 'home document'. in the generated html handout a fixed yellow button will
appear at the bottom of the document to provide a shortcut to the 'home document'.