add template for lecture notes / scriptums
1. Put extra content like learning outcomes, quizzes, ... in yaml block with meta data variables
2. Introduce a template for layout of lecture notes, using the meta data variables from (1)
3. Define a new default file to compile markdown lectures using the html template from (2)
4. Update html target in lecture makefile:
    a) Emit a new markdown using the lecture template, and
    b) pipe it into the normal process of markdown to html notes

This allows to change the structure of the html lecture notes more easily. Former lecture material
should be unaffected by this change, as it does not contain the new meta data variables and thus
step (4a) should produce the same markdown. Slides (beamer pdf) are not affected by this change.