improve shortcode img: allow arbitrary folder names (#19)
folder name for images was hard coded as 'images' for page-bundles or '<page>.images'
for single markdown pages. however, there is no need for this -- we just need to
concatenate the name of the pages folder to whatever the user specifies as src in
the image definition ...

note: we have a similar mechanism for attachments. however, there we need to keep this
since we automatically grab all .pdf files from the files-folder and thus we need to
distinguish the files-folder of different pages.