[HUGO] Lizenzangabe im Seitenfuß in Mobile-Version störend (#24)
* [Mobile] remove custom-comments.html

We used this partial to include the license file. However, it turned out, that
this will be kind of sticky. Using a mobile browser this would lead to a screen
half filled w/ content and half filled w/ license ...

* [Mobile] overwrite content-footer.html

Use partial content-footer.html to include the license statement, so it will
be at the end of the content and thus only be visible when scrolling to the end.

* Apply style in partial (not in license file)

* remove overwritten partial menu-footer

* re-add overwritten partial menu-footer

do not like the default, so let's use something customised ...

* Menu-Footer: use external file for links; use proper formatting

* fix typo

* fix typo

* improve formatting

* fix formatting